remove the hide on close trick made with windowmanager Fixes #453
